Combined extra machine controller file back into machine controller.
